# openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenQuester API",
    "version": "0.8.9"
  },
  "security": [
    {
      "SessionAuth": []
    }
  ],
  "paths": {
...
    "/v1/users/{id}": {
      "get": {
        "summary": "Get user info by id",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get user data successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUser"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update user by id",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputUpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update user data successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUser"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDeniedResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user by id",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDeniedResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/": {
      "get": {
        "summary": "Get all users info",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UsersSortBy"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationLimit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationOffset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all users data successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsers"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDeniedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Users not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "summary": "Get info about current user",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "Get user data successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUser"
                }
              }
            }
          },
          "401": {
            "description": "Invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update current user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputUpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update user data successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUser"
                }
              }
            }
          },
          "401": {
            "description": "Invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete current user",
        "tags": ["Users"],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "Invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session-based authentication using cookies."
      }
    },
    "schemas": {
      "PageInfo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "examples": [1000]
          }
        }
      },
...
      "ResponseUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "examples": [1]
          },
          "username": {
            "type": "string",
            "examples": ["John"]
          },
          "email": {
            "type": ["string", "null"],
            "examples": ["john.doe@gmail.com"]
          },
          "birthday": {
            "type": ["string", "null"],
            "format": "date",
            "examples": ["2024-01-01"]
          },
          "avatar": {
            "type": ["string", "null"],
            "description": "link on file GET",
            "examples": [
              "https://s3.asion.dev/openquester-sample-bucket/e/ec/ec65cd3756490723f8bcc3bc432b94a7"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "examples": ["2024-01-01T00:00:00Z"]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "examples": ["2024-01-01T00:00:00Z"]
          },
          "is_deleted": {
            "type": "boolean",
            "examples": [false]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "examples": [1]
                },
                "name": {
                  "type": "string",
                  "examples": ["get_all_users"]
                }
              }
            }
          }
        }
      },
...
      "InputUpdateUser": {
        "type": "object",
        "properties": {
          "username": {
            "type": ["string", "null"],
            "examples": ["John Doe"]
          },
          "email": {
            "type": ["string", "null"],
            "examples": ["user@example.com"],
            "format": "email"
          },
          "birthday": {
            "type": ["string", "null"],
            "format": "date",
            "examples": ["2024-01-01"]
          },
          "avatar": {
            "type": ["string", "null"],
            "examples": ["filename_of_file_in_bucket"]
          }
        }
      },
...
      "PaginatedUsers": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseUser"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
...
      "OrderDirection": {
        "type": "string",
        "enum": ["asc", "desc"],
        "examples": ["asc"],
        "default": "desc"
      },
      "PaginationLimit": {
        "type": "integer",
        "examples": [10],
        "minimum": 1,
        "maximum": 50,
        "default": 10
      },
      "PaginationOffset": {
        "type": "integer",
        "examples": [0],
        "minimum": 0
      },
...
      "UsersSortBy": {
        "type": ["string", "null"],
        "enum": ["id", "is_deleted", "created_at", "username", "email"],
        "examples": ["createdAt"],
        "default": "createdAt"
      },
...
      "InvalidSessionResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": ["Session invalid or expired"]
          }
        }
      },
      "UserNotFoundResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": ["User not found"]
          }
        }
      },
      "PermissionDeniedResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": ["You don't have permission to perform this action"]
          }
        }
      },
...
    }
  }
}